requirements.txt
django
flake8
pytest-django